fixed distribution error when redeeming tokens

+++ AskoLend-contracts/contracts/interfaces/MoneyMarketFactoryI.sol
@@ -11,10 +11,8 @@ The MoneyMarketFactoryI contract an abstract contract the MoneyMarketInstance us
     SafeMath instances
 **/
 
- abstract contract MoneyMarketFactoryI {
-
-
-   /**
+abstract contract MoneyMarketFactoryI {
+    /**
    @notice the createMMI function is used to initialize the MoneyMakerInstance and deploy its associated AHR && ALR token contracts
    @param _assetContractAdd is the address of the ERC20 asset being whitelisted
    @param _owner is the address that will own this contract(The AskoDAO)
@@ -23,37 +21,50 @@ The MoneyMarketFactoryI contract an abstract contract the MoneyMarketInstance us
    @dev this function uses ABI encoding to properly concatenate AHR- && ALR- in front of the tokens name and symbol
          before creating each token.
    **/
-     function createMMI(
-       address _assetContractAdd,
-       address _owner,
-       address _oracleFactory,
-       string memory _assetName,
-       string memory _assetSymbol
-     )
-     public
-     virtual
-     returns(address);
+    function createMMI(
+        address _assetContractAdd,
+        address _owner,
+        address _oracleFactory,
+        string memory _assetName,
+        string memory _assetSymbol
+    ) public virtual returns (address);
 
-     function liquidateBorrowAllowed(
-       address _lendedART,
-       address _collateralART,
-       address _liquidater,
-       address _borrower,
-       uint _repayAmount
-     )
-     public
-     virtual
-     returns(bool);
+    function liquidateBorrowAllowed(
+        address _lendedART,
+        address _collateralART,
+        address _liquidater,
+        address _borrower,
+        uint256 _repayAmount
+    ) public virtual returns (bool);
 
-     function checkAvailibleCollateralValue(address _borrower, address _ALR) external virtual returns(uint);
+    function checkAvailibleCollateralValue(address _borrower, address _ALR)
+        external
+        virtual
+        returns (uint256);
 
-    function trackCollateralUp(address _borrower, address _ALR, uint _amount) external virtual;
+    function trackCollateralUp(
+        address _borrower,
+        address _ALR,
+        uint256 _amount
+    ) external virtual;
 
-    function trackCollateralDown(address _borrower, address _ALR, uint _amount) external virtual;
+    function trackCollateralDown(
+        address _borrower,
+        address _ALR,
+        uint256 _amount
+    ) external virtual;
 
-    function lockCollateral(address _borrower, address _ALR, uint _amount) external virtual;
+    function lockCollateral(
+        address _borrower,
+        address _ALR,
+        uint256 _amount
+    ) external virtual;
 
-    function unlockCollateral(address _borrower, address _ALR, uint _amount) external virtual;
+    function unlockCollateral(
+        address _borrower,
+        address _ALR,
+        uint256 _amount
+    ) external virtual;
 
-    function _checkIfALR(address __inQ) external view virtual returns(bool);
+    function _checkIfALR(address __inQ) external view virtual returns (bool);
 }
